requirements.txt
# Core Dependencies
telebot==0.0.5
requests==2.31.0
Pillow==10.1.0
openai==1.12.0

# Hugging Face Ecosystem
huggingface-hub==0.19.4

# SSL/Networking
pyOpenSSL==23.3.0
urllib3==1.26.18

# Audio Processing (for future expansion)
pydub==0.25.1